corpuslio_django/requirements.txt
Django==5.0
djangorestframework==3.14.0
drf-spectacular==0.29.0
celery>=5.2.7
redis>=4.3.4
# Database
psycopg2-binary>=2.9.3
dj-database-url>=2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
django-ratelimit==4.1.0

# Static serving for simple deployments
whitenoise==6.4.0

# Existing OCRchestra dependencies
jsonschema
pytest
requests==2.31.0
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.1.0
sqlalchemy
groq
pdfplumber==0.10.3
pandas
plotly

# Export formats
reportlab==4.0.7
openpyxl==3.1.2
django-allauth
